Consolidate component modules
