+++ src/components/TileDescription.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
 import { FaClipboardList } from 'react-icons/fa';
-import Description from './Description';
 import MiniButton from './MiniButton';
 
 const FooterContainer = styled.div`
@@ -27,6 +26,34 @@ const ButtonContainer = styled.div`
     margin-right: 3px;
     margin-bottom: 13px;
 `;
+
+const DescriptionContainer = styled.div`
+    text-align: left;
+    margin-top: 0px;
+    padding-left: 10px;
+    padding-right: 10px;
+
+    h3 {
+        text-align: center;
+    }
+
+    ul {
+        display: grid;
+        grid-template-columns: 1fr 1fr;
+    }
+`;
+
+const Description = props => (
+    <DescriptionContainer>
+        <p>{props.description}</p>
+        <h3>Built using</h3>
+        <ul>
+            {props.builtUsing.map(tech => (
+                <li key={`${props.title}-${tech}`}>{tech}</li>
+            ))}
+        </ul>
+    </DescriptionContainer>
+)
 
 const TileDescription = props => {
     const proj = props.project;
